Enhance generate-sources.sh for tarball creation
Updated the script to handle both release and snapshot cases for generating tarballs.

+++ server/generate-sources.sh
@@ -2,10 +2,18 @@
 set -xeu
 NAME=postgresql
 : ${SOURCE_VERSION:?The SOURCE_VERSION environment variable is required}
-WORKDIR=$(pwd)                             
+WORKDIR=$(pwd)
 TARNAME="${NAME}-${SOURCE_VERSION}"
 cd ${WORKDIR}
-cp -r src/ "${TARNAME}"                   
-tar -cjf "${TARNAME}.tar.bz2" "${TARNAME}/"  
-md5sum "${TARNAME}.tar.bz2" > "${TARNAME}.tar.bz2.md5"
-rm -rf "${TARNAME}"                        
+
+if [ -n "${URL:-}" ]; then
+  # release case: download tarball from URL
+  wget -O "${TARNAME}.tar.bz2" ${URL}
+  md5sum "${TARNAME}.tar.bz2" > "${TARNAME}.tar.bz2.md5"
+else
+  # snapshot case: create tarball from git checkout
+  cp -r src/ "${TARNAME}"
+  tar -cjf "${TARNAME}.tar.bz2" "${TARNAME}/"
+  md5sum "${TARNAME}.tar.bz2" > "${TARNAME}.tar.bz2.md5"
+  rm -rf "${TARNAME}"
+fi
